Learned a new way of using ranges. You cannot create a slice out of the blue like in Python. Either allocate a vector (or array when the container size is known in advance), or use the Range / RangeInclusive, maybe omitting type specifications and letting the compiler infer them.
Also see this:
https://stackoverflow.com/questions/29500666/what-is-the-difference-between-len-and-count
Seems like ranges cannot be used because they do not have len() method,
therefore we have to use slices. But slices appear from vectors, so what
can I do then?

Seems like there is no answer. I cannot use slice for non-allocated
memory => need to use regular integers like in C++.

1) Use a fake container just to satisfy Rust.
2) Use plain numbers - this is less preferable, as you can easily make
mistakes.

--- A small note about binary search
Need to always include 2 checks if we are looking for the number that is
closest to the one being searched for. See play/src/main.rs. Also need
only one number in case we are looking for an element that is for sure
present in the container.